+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## 12.3.0
 
-* Serve the [OpenAPI specification](https://www.openapis.org/) under the API route `/swagger/`
+* Serve the [OpenAPI specification](https://www.openapis.org/) under the API route `/spec/`
 
 ## 12.2.0